typegraph/pyproject.toml
[tool.poetry]
name = "typegraph"
version = "0.0.3-dev.3"
description = "Free and open ecosystem for API composition."
authors = ["Metatype Contributors <support@metatype.dev>"]
license = "ELv2"
homepage = "https://metatype.dev"
repository = "https://github.com/metatypedev/metatype"
include = ["typegraph/**/*", "LICENSE.md"]
keywords = ["api", "composition", "typesystem", "graphql", "ecosystem"]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python",
]

[tool.poetry.scripts]
py-tg = "typegraph.utils.loaders:cmd"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
httpx = { extras = ["http2"], version = ">=0.22,<0.24" }
redbaron = "^0.9.2"
frozendict = "2.3.4"
semver = "^2.13.0"
PyYAML = "^6.0"
deepmerge = "^1.1.0"
graphql-core = "^3.2.3"
gql = { extras = ["requests"], version = "^3.4.0" }
ordered-set = "^4.1.0"
typing-extensions = "^4.4.0"
attrs = "^22.2.0"
astunparse = "^1.6.3"
black = ">=22.12,<24.0"
strenum = "^0.4.9"
python-box = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-snapshot = "^0.9.0"
pydoc-markdown = { git = "https://github.com/metatypedev/pydoc-markdown", branch = "develop" }
ruff = "^0.0.245"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
